Created MacNotesViewController to separate exclusive properties and methods
